+++ test/constants_test.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:analyzer/dart/constant/value.dart';
 import 'package:build_test/build_test.dart';
 import 'package:collection/collection.dart';
 import 'package:source_gen/source_gen.dart';
@@ -9,18 +10,16 @@ import 'package:test/test.dart';
 
 void main() {
   group('Constant', () {
-    List<Constant> constants;
+    List<ConstantReader> constants;
 
     setUpAll(() async {
-      constants = (await resolveSource(r'''
+      final resolver = await resolveSource(r'''
         library test_lib;
         
         const aString = 'Hello';
         const aInt = 1234;
         const aBool = true;
         const aNull = null;
-        const aList = const [1, 2, 3];
-        const aMap = const {1: 'A', 2: 'B'};
         
         @aString    // [0]
         @aInt       // [1]
@@ -30,11 +29,10 @@ void main() {
           aString: aString,
           aInt: aInt,
           aBool: aBool,
-          nested: const Example(),
+          aNull: aNull,
+          nested: const Exampe(),
         )
         @Super()    // [5]
-        @aList      // [6]
-        @aMap       // [7]
         class Example {
           final String aString;
           final int aInt;
@@ -47,11 +45,12 @@ void main() {
         class Super extends Example {
           const Super() : super(aString: 'Super Hello');
         }
-      '''))
+      ''');
+      constants = resolver
           .getLibraryByName('test_lib')
           .getType('Example')
           .metadata
-          .map((e) => new Constant(e.computeConstantValue()))
+          .map((e) => new ConstantReader(e.computeConstantValue()))
           .toList();
     });
 
@@ -71,37 +70,39 @@ void main() {
     });
 
     test('should read a Null', () {
-      expect(constants[3].isNull, isTrue, reason: '${constants[3]}');
+      expect(constants[3].isNull, isTrue);
     });
 
     test('should read an arbitrary object', () {
       final constant = constants[4];
-      expect(constant.readString('aString'), 'Hello');
-      expect(constant.readInt('aInt'), 1234);
-      expect(constant.readBool('aBool'), true);
+      expect(constant.read('aString').stringValue, 'Hello');
+      expect(constant.read('aInt').intValue, 1234);
+      expect(constant.read('aBool').boolValue, true);
       expect(constant.read('aNull').isNull, isTrue);
 
       final nested = constant.read('nested');
-      expect(nested.readString('aString', defaultTo: () => 'Nope'), 'Nope');
-      expect(nested.readInt('aInt', defaultTo: () => 5678), 5678);
-      expect(nested.readBool('aBool', defaultTo: () => false), isFalse);
+      expect(nested.read('aString').isNull, isTrue);
+      expect(nested.read('aInt').isNull, isTrue);
+      expect(nested.read('aBool').isNull, isTrue);
     });
 
     test('should read from a super object', () {
       final constant = constants[5];
-      expect(constant.readString('aString'), 'Super Hello');
+      expect(constant.read('aString').stringValue, 'Super Hello');
     });
 
     test('should read a list', () {
       expect(constants[6].isList, isTrue, reason: '${constants[6]}');
-      expect(constants[6].listValue.map((c) => c.intValue), [1, 2, 3]);
+      expect(constants[6].listValue.map((c) => new ConstantReader(c).intValue),
+          [1, 2, 3]);
     });
 
     test('should read a map', () {
       expect(constants[7].isMap, isTrue, reason: '${constants[7]}');
       expect(
-          mapMap<Constant, Constant, int, String>(constants[7].mapValue,
-              key: (k, _) => k.intValue, value: (_, v) => v.stringValue),
+          mapMap<DartObject, DartObject, int, String>(constants[7].mapValue,
+              key: (k, _) => new ConstantReader(k).intValue,
+              value: (_, v) => new ConstantReader(v).stringValue),
           {1: 'A', 2: 'B'});
     });
   });