+++ CHANGELOG.md
@@ -33,14 +33,54 @@
   ```
   
   In Dart SDK `>=1.25.0` this can be relaxed as `part of` can refer to a path.
+  To opt-in, `GeneratorBuilder` now has a new flag, `requireLibraryDirective`.
+  Set it to `false`, and also set your `sdk` constraint appropriately:
+  
+  ```yaml
+    sdk: '>=1.25.0 <2.0.0'
+  ```
 
-* Added `findType`, an utility method for `LibraryElement#getType` that also
-  traverses `export` directives for publicly exported types. For example, to
-  find `Generator` from `package:source_gen/source_gen.dart`:
+* Added `LibraryReader`, a utility class for `LibraryElement` that exposes
+  high-level APIs, including `findType`, which traverses `export` directives
+  for publicly exported types. For example, to find `Generator` from
+  `package:source_gen/source_gen.dart`:
   
   ```dart
   void example(LibraryElement pkgSourceGen) {
-    findType(pkgSourceGen, 'Generator');
+    var library = new LibraryReader(pkgSourceGen);
+  
+    // Instead of pkgSourceGen.getType('Generator'), which is null.
+    library.findType('Generator');
+  }
+  ```
+
+* Added `ConstantReader`, a high-level API for reading from constant (static)
+  values from Dart source code (usually represented by `DartObject` from the
+  `analyzer` package):
+  
+  ```dart
+  abstract class ConstantReader {
+    factory ConstantReader(DartObject object) => ...
+  
+    // Other methods and properties also exist.
+  
+    /// Reads[ field] from the constant as another constant value.
+    ConstantReader read(String field);
+  
+    /// Reads [field] from the constant as a boolean.
+    ///
+    /// If the resulting value is `null`, uses [defaultTo] if defined.
+    bool readBool(String field, {bool defaultTo()});
+  
+    /// Reads [field] from the constant as an int.
+    ///
+    /// If the resulting value is `null`, uses [defaultTo] if defined.
+    int readInt(String field, {int defaultTo()});
+  
+    /// Reads [field] from the constant as a string.
+    ///
+    /// If the resulting value is `null`, uses [defaultTo] if defined.
+    String readString(String field, {String defaultTo()});
   }
   ```